+++ Sources/KeyboardKit/_Deprecated/KeyboardUrlOpener.swift
@@ -11,12 +11,12 @@ import Foundation
 /**
  This class can be used to open URLs from a keyboard without
  having to use `UIApplication`.
-
- You can use ``KeyboardUrlOpener/shared`` to avoid having to
- create custom instances. Note that you have to manually set
- the ``controller`` when you create a custom opener or use a
- custom controller.
+ 
+ This is deprecated in KeyboardKit 8.2. Instead of using the
+ shared instance, use the controller openUrl directly or the
+ action handler, by triggering ``KeyboardAction/url(_:id:)``.
  */
+@available(*, deprecated, message: "This class is deprecated. Either use the controller directly, or trigger a KeyboardAction.url to handle URLs with the action handler.")
 public class KeyboardUrlOpener {
 
     public init() {}
@@ -29,11 +29,20 @@ public class KeyboardUrlOpener {
     public static let shared = KeyboardUrlOpener()
 
     /// The controller to use to open the URL.
-    public unowned var controller: KeyboardController?
+    public var controller: KeyboardController? {
+        get { KeyboardUrlOpenerInternal.controller }
+        set { KeyboardUrlOpenerInternal.controller = newValue
+        }
+    }
 
     /// Open a custom URL.
     public func open(_ url: URL?) throws {
         guard let controller else { throw UrlError.noKeyboardController }
         controller.openUrl(url)
     }
+}
+
+class KeyboardUrlOpenerInternal {
+    
+    public weak static var controller: KeyboardController?
 }