+++ Sources/KeyboardKit/_Keyboard/KeyboardInputViewController.swift
@@ -28,10 +28,6 @@ open class KeyboardInputViewController: UIInputViewController, KeyboardControlle
 
 
     // MARK: - View Controller Lifecycle
-    
-    deinit {
-        viewWillUnregisterSharedController()
-    }
 
     open override func viewDidLoad() {
         super.viewDidLoad()
@@ -76,23 +72,14 @@ open class KeyboardInputViewController: UIInputViewController, KeyboardControlle
             }
         }
     }
-    
+
     /**
      This function is called when the controller is about to
      register itself as the shared controller.
      */
     open func viewWillRegisterSharedController() {
         NextKeyboardController.shared = self
-        KeyboardUrlOpener.shared.controller = self
-    }
-    
-    /**
-     This function is called when the controller is about to
-     deinit, to unregister itself as the shared controller.
-     */
-    open func viewWillUnregisterSharedController() {
-        NextKeyboardController.shared = nil
-        KeyboardUrlOpener.shared.controller = nil
+        KeyboardUrlOpenerInternal.controller = self
     }
 
     /**